requirements.txt
bottle==0.13.2
branca==0.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clr_loader==0.2.7.post0
distlib==0.3.9
docutils==0.21.2
et_xmlfile==2.0.0
feedparser==6.0.11
filelock==3.16.1
folium==0.19.4
geopandas==1.0.1
idna==3.10
Jinja2==3.1.5
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivy-garden.mapview==1.0.6
mapview==1.0.6
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
proxy_tools==0.1.0
pycparser==2.22
Pygments==2.18.0
pyogrio==0.10.0
pypiwin32==223
pyproj==3.7.0
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2024.2
pywebview==5.4
pywin32==308
requests==2.32.3
sgmllib3k==1.0.0
shapely==2.0.7
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
xyzservices==2025.1.0
